% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_state.R
\name{.complete_states}
\alias{.complete_states}
\title{Fill in missing rows using enter and exit functions}
\usage{
.complete_states(from, to, id, enter, exit, max_id)
}
\arguments{
\item{from, to}{Data.frames to tween between}

\item{id}{The name of the column that holds the matching id}

\item{enter, exit}{functions to fill out missing rows in \code{from} and \code{to}
respectively}
}
\value{
A list with the elements \code{from} and \code{to} holding the filled out
versions of \code{from} and \code{to}
}
\description{
This function figures out which rows are missing in either state and applies
the provided \code{enter} and \code{exit}  functions to fill in the blanks and provide
a 1-to-1 relation between the rows in \code{from} and \code{to}.
}
\keyword{internal}
