% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_state.R
\name{.max_id}
\alias{.max_id}
\title{Get the highest id occuring in a dataset}
\usage{
.max_id(data)
}
\arguments{
\item{data}{A data.frame as returned by \code{tween_state}}
}
\value{
An integer giving the currently highest id
}
\description{
This is helper for \code{tween_state} related functions to get the currently
highest \code{.id} in a frame collection
}
\keyword{internal}
