% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_keyframe.R
\name{interpolate_custom_at}
\alias{interpolate_custom_at}
\title{Fallback for keyframe vector support}
\usage{
interpolate_custom_at(from, to, at, ease)
}
\arguments{
\item{from, to}{vectors to interpolate between}

\item{at}{value between 0 and 1 defining the point}

\item{ease}{the easing function to use}
}
\description{
Fallback for keyframe vector support
}
\keyword{internal}
