% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_keyframe.R
\name{prepare_keyframes}
\alias{prepare_keyframes}
\title{Prepare keyframes for generator}
\usage{
prepare_keyframes(.data, keyframe)
}
\arguments{
\item{.data}{A keyframe generator}

\item{keyframe}{A keyframe to add}
}
\value{
A valid keyframe generator
}
\description{
Prepare keyframes for generator
}
\keyword{internal}
