% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_states.R
\name{tween_states}
\alias{tween_states}
\title{Tween a list of data.frames representing states}
\usage{
tween_states(data, tweenlength, statelength, ease, nframes)
}
\arguments{
\item{data}{A list of data.frames. Each data.frame must contain the same
number of rows, but only the first data.frame needs to contain all columns.
Subsequent data.frames need only contain the columns that shows change.}

\item{tweenlength}{The lengths of the transitions between each state.}

\item{statelength}{The length of the pause at each state.}

\item{ease}{The easing functions to use for the transitions. See details.}

\item{nframes}{The number of frames to generate. The actual number of frames
might end up being higher depending on the regularity of \code{tweenlength}
and \code{statelength}.}
}
\value{
A data.frame with the same columns as the first data.frame in
\code{data}, but replicated \code{nframes} times. An additional column called
\code{.frame} will be added giving the frame number.
}
\description{
This function is intended to create smooth transitions between states of
data. States are defined as full data.frames or data.frames containing only
the columns with change. Each state can have a defined period of pause, the
transition length between each states can be defined as well as the easing
function.
}
\details{
How transitions proceed between states are defined by an easing function. The
easing function converts the parameterized progression from one state to the
next to a new number between 0 and 1. \code{linear} easing is equivalent to
an identity function that returns the input unchanged. In addition there are
a range of additional easers available, each with three modifiers.

\strong{Easing modifiers:}
\describe{
 \item{-in}{The easing function is applied as-is}
 \item{-out}{The easing function is applied in reverse}
 \item{-in-out}{The first half of the transition it is applied as-is, while
 in the last half it is reversed}
}

\strong{Easing functions}
\describe{
 \item{quadratic}{Models a power-of-2 function}
 \item{cubic}{Models a power-of-3 function}
 \item{quartic}{Models a power-of-4 function}
 \item{quintic}{Models a power-of-5 function}
 \item{sine}{Models a sine function}
 \item{circular}{Models a pi/2 circle arc}
 \item{exponential}{Models an exponential function}
 \item{elastic}{Models an elastic release of energy}
 \item{back}{Models a pullback and relase}
 \item{bounce}{Models the bouncing of a ball}
}

A good visual explanation can be found \href{http://easings.net}{here}.
}

