% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{vec_tween_class}
\alias{vec_tween_class}
\title{Get the nominal class of a vector}
\usage{
vec_tween_class(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Get the nominal class of a vector
}
\keyword{internal}
